Add welcome section and factory/module functions
The welcome section allows for user input: displaying
the name, what player they are, along with which symbol.
No more than 2 can be added.
This can be removed by clicking on the svg.

The module GameBoard will hold the state of the game in an array.
Current functionality is only adding a player to the game.
The player factory function at this point is mostly useless as
I am trying to figure out how to properly connect everything cleanly.