Make displayController handle display and gameFlow handle turns.
DisplayController will self execute and add an eventlistener
on the 9 div elements that make up the game board.
Any click will emit 'moveMade' indicating a move has been made
by a player. updateBoard is listening for this and will change
the gameBoard internally to the correct move with the help of gameFlow.
Then updateBoard will emit 'updateBoardDisplay' indicating that
the board has been changed and the move needs to be displayed.
displayController will read the gameBoard down the list and place
the corresponding svg elements in the divs.

gameFlow simply listens for when the baord is updated and
changes turn to whoevers turn it is not. It emits the correct marker
for the currents persons turn, so that updateBoard in GameBoard places
the correct marker internally.
Also added is a start/restart button functionality
that prevents users from placing ANY marks
until it is clicked, with there being two players added.
The appropiate functions have a boolean variable that facilitates this.
Which has been allowed to change when 'gameOn' is emitted.