Add a bot that will make random moves
Make a button that will add a bot 'player'
to the list of players and set the isBot property
to be true.

When a move is made, the bot controller module will emit that
it is now a bot's turn. gameboard will receive
this and no longer allow the players' clicks on the board
to be registered as a move.
The module will also emit 'botMoveMade' so that the gameboard
can call updateBoardFromBot.
This is to keep track of whether a move was made from a player click,
or a bot move.

Include certain lines that allow for two bots to 'play' eachother
---------------------------------------------------------------------------
Add function to indicate who's turn it